+++ source/patterns/@aws-solutions-constructs/aws-events-rule-kinesisfirehose-s3/test/events-rule-kinesisfirehose-s3.test.ts
@@ -1,0 +1,127 @@
+/**
+ *  Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance
+ *  with the License. A copy of the License is located at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  or in the 'license' file accompanying this file. This file is distributed on an 'AS IS' BASIS, WITHOUT WARRANTIES
+ *  OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions
+ *  and limitations under the License.
+ */
+
+import { SynthUtils } from '@aws-cdk/assert';
+import * as cdk from "@aws-cdk/core";
+import * as events from "@aws-cdk/aws-events";
+import '@aws-cdk/assert/jest';
+import {EventsRuleToKinesisFirehoseToS3, EventsRuleToKinesisFirehoseToS3Props} from '../lib';
+
+// --------------------------------------------------------------
+// Test snapshot match with default parameters
+// --------------------------------------------------------------
+function deployNewStack(stack: cdk.Stack) {
+  const props: EventsRuleToKinesisFirehoseToS3Props = {
+    eventRuleProps: {
+      description: 'event rule props',
+      schedule: events.Schedule.rate(cdk.Duration.minutes(5))
+    }
+  };
+  return new EventsRuleToKinesisFirehoseToS3(stack, 'test-events-rule-kinesis-firehose-s3-default-parameters', props);
+}
+
+test('Test snapshot match with default parameters', () => {
+  const stack = new cdk.Stack();
+  deployNewStack(stack);
+
+  // Assertions
+  expect(SynthUtils.toCloudFormation(stack)).toMatchSnapshot();
+});
+
+// --------------------------------------------------------------
+// Test properties
+// --------------------------------------------------------------
+test('Test properties', () => {
+  const stack = new cdk.Stack();
+  const construct: EventsRuleToKinesisFirehoseToS3 = deployNewStack(stack);
+
+  // Assertions
+  expect(construct.eventsRule !== null);
+  expect(construct.eventsRole !== null);
+  expect(construct.kinesisFirehose !== null);
+  expect(construct.kinesisFirehoseRole !== null);
+  expect(construct.kinesisFirehoseLogGroup !== null);
+  expect(construct.s3Bucket !== null);
+  expect(construct.s3LoggingBucket !== null);
+  expect(construct.firehoseToS3 !== null);
+});
+
+// --------------------------------------------------------------
+// Test default server side s3 bucket encryption
+// --------------------------------------------------------------
+test('Test default server side s3 bucket encryption', () => {
+  const stack = new cdk.Stack();
+  deployNewStack(stack);
+
+   // Assertions
+  expect(stack).toHaveResource('AWS::S3::Bucket', {
+    BucketEncryption: {
+      ServerSideEncryptionConfiguration: [
+        {
+          ServerSideEncryptionByDefault: {
+            SSEAlgorithm: "AES256"
+          }
+        }
+      ]
+    }
+  });
+});
+
+// --------------------------------------------------------------
+// Test property override
+// --------------------------------------------------------------
+test('Test property override', () => {
+  const stack = new cdk.Stack();
+
+  // create properties
+  const props: EventsRuleToKinesisFirehoseToS3Props = {
+    eventRuleProps: {
+      description: 'event rule props',
+      schedule: events.Schedule.rate(cdk.Duration.minutes(5))
+    },
+    kinesisFirehoseProps: {
+      extendedS3DestinationConfiguration: {
+        bufferingHints: {
+          intervalInSeconds: 600,
+          sizeInMBs: 55
+        },
+      }
+    },
+    bucketProps: {
+      blockPublicAccess: {
+        blockPublicAcls: false,
+        blockPublicPolicy: true,
+        ignorePublicAcls: false,
+        restrictPublicBuckets: true
+      }
+    }
+  };
+  new EventsRuleToKinesisFirehoseToS3(stack, 'test-events-rule-firehose-s3', props);
+
+  expect(stack).toHaveResource("AWS::S3::Bucket", {
+    PublicAccessBlockConfiguration: {
+      BlockPublicAcls: false,
+      BlockPublicPolicy: true,
+      IgnorePublicAcls: false,
+      RestrictPublicBuckets: true
+    },
+  });
+
+  expect(stack).toHaveResourceLike("AWS::KinesisFirehose::DeliveryStream", {
+    ExtendedS3DestinationConfiguration: {
+      BufferingHints: {
+        IntervalInSeconds: 600,
+        SizeInMBs: 55
+      }
+  }});
+});